Move shared_helpers from gold
