+++ source/core/src/core/utility/shared_helpers.py
