+++ source/core/tests/gold/infrastructure/adapters/test_delta_gold_adapter.py
@@ -10,7 +10,7 @@ from core.settings.catalog_settings import CatalogSettings
 
 
 @mock.patch("os.getenv")
-@mock.patch("core.gold.infrastructure.shared_helpers.get_storage_base_path")
+@mock.patch("core.utility.shared_helpers.get_storage_base_path")
 def test__start_write_stream__should_write_to_gold_table(
     mock_get_checkpoint_path, mock_getenv, spark: SparkSession, migrations_executed
 ):
